Day 8 Part 2: Check visible_distance
To maybe make the poking around a little easier, we add a Point type
with i32 indexes, so as we scan we can just repeatedly add another Point
as a movement vector.

Point also has a get into a Array2, which checks if the index is
negative and returns None, otherwise delegating to the Array2 to check
for positive bounds, and maybe return the height of the tree there.