Day 11 Part 2: limiting the values
For part 2 we need to iterate 10_000 times without dividing by 3
each check.  This makes the numbers unmanagable, they don't even fit
into a u128.   We could try for brute force and use a bigint library,
but that's just overkill (plus too slow when I spiked it)

Instead we can calculate a common multiple of all the monkeys test
values, and then use to limit the values we pass around by taking the
remainder of value/limit.

We could possibly fit this back into a u32 if we computed the lcm rather
than simply the product of all values, but the product is much simpler
to calculate.