Day 12 Part 2: Brute force
We just find all the starts, then all the shortest paths from them.

The better way would be to fix the neighbours function to invert the
filter, so we find the shortest path from end to start, and then call
dijkstra_all starting at and, and extract all the start nodes.