Day 15 Part 1: Bruteish computation of excluded_cells
This is a little brute force.  We compute a few sets of ranges, and then
punch them into a HashSet just so we can count the entries.  For the small
example this is fine, but it's almost a second in runtime for the actual
solve.

        Running `target/release/15`
    Part 1: 4717631 (894.3ms)

More efficient would be to actually maintain a set of ranges, merging
up, and then splitting out the actual beacons.  Will see what part 2
needs first.