Day 15 Part 2: Guided bruteforce
For each of the y values we're scanning, generate RangeSet of that y
(check_xs, extracted from the part one solution excluded_cells).
Take the intersection of that RangeSet with 0..=dim, and then check if
the intersection is 2 ranges.  If it's 2 ranges then there must be
our gap right between them, which is what we're searching for.

Runtime isn't too great:

    Part 2: 13197439355220 (2.3s)

Might be better if we culled during the call to scanned_xs