Day 16, Part 1: BFS over a reduced network
We use the Floyd-Warshall algorithm to determine all the shortest paths
in the graph between two nodes.

Then we extract the `interesting` nodes to identify all the valves that
will do something, and do a BFS reach from 'AA' to all 'interesting'
nodes, using the distances array to calculate the steam flow for this
walk.

We use a bitmap from the Bitmaps crate to track which of the interesting
nodes are still closed, because it hashes so can used for the state
handled by bfs_reach.

Bashed out head against a brick wall for quite a while before realising
that we'd assumed valve[0] == valve.AA.  That's true for the example,
but *not* for the real data.