Day 17: Build a self-playing Tetris
This one was fun enough to code up.

We have a Well structure, that has a vec of u8 for each row in the well.

The rock shapes all fit into a u32, and we get to do some bit-masking to
see if we have collided with the wall or existing blocks.

There's a little ugliness in how we ensure there's enough space in the
well for our next shape, maybe we should just preallocate and then as we
draw we can skip down until we hit a y we like.