Day 17 Part 2: Classic 'find cycle skip'
This is a typical AOC complication, given a simulation run it for a
ridiculous number of cycles.

Here having the well as a Vec<u8> really works for us, as we can just
take the top COUNT bytes of the well (last COUNT of the vec) as our
fingerprint for finding the repeat.