Day 20 Part 1:  Implement mix
Track the indexes, then for each value, pull out that index and reinsert
it.

Feels like it could be somewhat inefficient, as we're removing from the
middle of a vec and then reinserting, as that's two memmoves of at worst
the entire array.  The dataset is small enough that it's not worth doing
a doubly linked list (or simulation with arrays).

Small optimisation for if we're inserting at the front, it's the
same as adding to the end in this usecase, so we know we'll avoid some
memcpys on that path.