Refactor tree to be a real tree.
For part 2 we can find our input variable by identifying the parent of
'root' with humn as an ancestor, and the one without.  The one without
gives the contstant we want to get to, so we could walk the tree
backwards up to humn computing the inverse of each step.

We didn't build our tree really as a tree though, so some of the
places we'd want to walk back up were getting lost in lack of helpers.

This commit changes to a tree rooted at the 'root' element.  We tossed
out nom as it wasn't helping that much as we needed two passes over the
data, once to break up the names from thier definitions, and then a
second to recursively turn the expressions into nested tree nodes.