Day 21 Part 2: Uneval the tree.
Now we have a constant folded tree, we can check which of the left/right
is folded (becomes Monkey::Value(_)) and repeatedly walk up the other
leg inverting all the operations on accumulator (`uneval`) as as we go.

Once we reach the humn / Monkey::Variable, what's left is the answer.