winnow: finish porting
One neat thing from winnow is it blanket implements Parser for `char`
and `str`, so we don't need to use `literal`/`tag`.

Also throw out the Stream type alias cargo.  From reading the tutorial
it becomes more apparent taht the input to a parser should be a mut
pointer to something, which is why the signature no longer returns the
leftover, it just advances the pointer when accepting.
`input: &mut &str` looks a little ugh, but it isn't too bad.