Day 22 Part 2: Implement the CubeNet zooper
Needed paper, a rubik's cube, a sharpie, and lots of talking to myself
to encode both the nets, and the CubeNet Zooper implementation.

Made some use of tarpaulin to get a coverage report to flush out the
bugs in the CubeNet Zooper trait, as there are a lot of caces to verify.

Tarpaulin default to deleting your target dir because it has trust issues
around recompilation not tracking RUSTFLAGS correctly.   To avoid constant
rebuilds, give it its own target-dir eg:

    cargo tarpaulin --target-dir tarp-target --bin 22 --out html